+++ packages/models/src/constants/editor.ts
@@ -55,18 +55,6 @@ export enum Theme {
     System = 'system',
 }
 
-export enum Language {
-    English = 'en',
-    Japanese = 'ja',
-    Chinese = 'zh',
-}
-
-export const LANGUAGE_DISPLAY_NAMES: Record<Language, string> = {
-    [Language.English]: 'English',
-    [Language.Japanese]: '日本語',
-    [Language.Chinese]: '中文',
-} as const;
-
 export const DefaultSettings = {
     SCALE: 0.7,
     PAN_POSITION: { x: 175, y: 100 },
