+++ packages/models/src/constants/index.ts
@@ -3,4 +3,5 @@ export * from './dom';
 export * from './editor';
 export * from './freestyle';
 export * from './ipc';
+export * from './language';
 export * from './terminal';
